feat: support export all

+++ src/mcr_analyzer/utils/number.py
@@ -1,2 +1,2 @@
-def clamp(*, x: int, lower_bound: int, upper_bound: int) -> int:
+def clamp(*, x: float, lower_bound: float, upper_bound: float) -> float:
     return max(lower_bound, min(x, upper_bound))
